app/requirements.txt
# API
fastapi
uvicorn
# データベース

# その他
dotenv

# LLM
langchain_openai
langchain_core
langchain_community
pydantic